pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tokens"
version = "0.1.0"
description = "Code for token-level representation alignment experiments from The Truth Lies Somewhere in the Middle (of the Generated Tokens)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "accelerate",
  "datasets",
  "huggingface-hub==0.30.2",
  "numpy",
  "pandas",
  "pillow",
  "safetensors==0.5.3",
  "torch",
  "torchvision",
  "tokenizers==0.21.1",
  "tqdm",
  "transformers==4.51.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
tokens-embed = "tokens.embed:main"
tokens-tokenwise = "tokens.tokenwise:main"
